# docs/lifecycle/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "robotsix-central-deploy Lifecycle API",
    "description": "Start, stop, restart, and inspect suite services.",
    "version": "0.1.0",
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/chat/github/repos/{owner}/{repo}": {
      "get": {
        "tags": [
          "chat-github"
        ],
        "summary": "Get GitHub repository details",
        "description": "Get *owner*/*repo*'s details (visibility, description, settings).",
        "operationId": "get_repo_chat_github_repos__owner___repo__get",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Repo Chat Github Repos  Owner   Repo  Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Repository not found or App not installed on it"
          },
          "503": {
            "description": "GitHub App not configured"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "chat-github"
        ],
        "summary": "Update GitHub repository settings",
        "description": "Update *owner*/*repo*'s settings \u2014 only provided fields are changed.",
        "operationId": "update_repo_chat_github_repos__owner___repo__patch",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRepoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Repo Chat Github Repos  Owner   Repo  Patch"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Repository not found or App not installed on it"
          },
          "422": {
            "description": "No fields provided, or GitHub rejected the request"
          },
          "503": {
            "description": "GitHub App not configured"
          }
        }
      }
    },
    "/chat/github/repos": {
      "post": {
        "tags": [
          "chat-github"
        ],
        "summary": "Create a new GitHub repository",
        "description": "Create a new repository under the configured account.\n\nUses ``github_repo_create_token`` (a PAT), not the GitHub App\ninstallation token \u2014 GitHub Apps cannot create repositories under a\npersonal account. The repository is always created under that token's\nown account; there is no way to target an arbitrary owner.\n\nAlso enables the Dependency graph / Dependabot alerts (see\n:func:`enable_vulnerability_alerts`) so CI checks that depend on it\n(e.g. ``dependency-review-action``) don't fail on day one.",
        "operationId": "create_repo_chat_github_repos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRepoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Repo Chat Github Repos Post"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Repository already exists"
          },
          "422": {
            "description": "GitHub rejected the request (e.g. invalid name)"
          },
          "503": {
            "description": "Repo creation not configured"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRepoRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Repository name (no owner prefix)."
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Repository description.",
            "default": ""
          },
          "private": {
            "type": "boolean",
            "title": "Private",
            "description": "Create as a private repository.",
            "default": false
          },
          "homepage": {
            "type": "string",
            "title": "Homepage",
            "description": "Homepage URL.",
            "default": ""
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topics",
            "description": "Topics to attach after creation."
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateRepoRequest",
        "description": "Body for ``POST /chat/github/repos``."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateRepoRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "New description."
          },
          "private": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Private",
            "description": "New visibility."
          },
          "has_issues": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Issues",
            "description": "Enable/disable Issues."
          },
          "has_wiki": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Wiki",
            "description": "Enable/disable the Wiki."
          },
          "allow_auto_merge": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow Auto Merge",
            "description": "Allow auto-merge on pull requests."
          },
          "delete_branch_on_merge": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delete Branch On Merge",
            "description": "Automatically delete head branches after merge."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UpdateRepoRequest",
        "description": "Body for ``PATCH /chat/github/repos/{owner}/{repo}``.\n\nAll fields are optional; only the ones provided are changed.\nUnknown keys are rejected with 422."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
  "security": [],
  "servers": [
    {
      "url": "http://localhost:8100",
      "description": "Local development server"
    }
  ]
}